backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.5.3
pydantic-settings==2.1.0
pandas==2.2.1
nest-asyncio==1.6.0
websockets==12.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
aiofiles==23.2.1
